BUGFIX THRIFT: readDoulbe used parseFloat to parse a float value which results in bad values. Updated thrift files too.

+++ Collector/src/de/tum/in/sonar/collector/tsdb/MetricPoint.java
@@ -21,7 +21,7 @@ public class MetricPoint {
 		this.value = value;
 	}
 
-	final long getValue() {
+	final double getValue() {
 		return value.getValue();
 	}
